Add form style and nav
