+++ src/components/Layout.module.css
@@ -1,0 +1,5 @@
+.main {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
