+++ src/index.css
@@ -1,4 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@400;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Arizonia&display=swap');
 
 * {
   box-sizing: border-box;
@@ -12,6 +13,10 @@ html {
 
 .form-control {
   margin-bottom: 1rem;
+}
+
+a {
+  text-decoration: none;
 }
 
 .form-control input,
@@ -36,8 +41,7 @@ html {
 
 .form-control input:focus {
   outline: none;
-  border-color: #240370;
-  background-color: #e0d4fd;
+  background-color: rgba(152, 192, 31, 0.2);
 }
 
 .control-group {
@@ -53,9 +57,9 @@ html {
 
 button {
   font: inherit;
-  background-color: #240370;
+  background-color: #97BF1C;
   color: white;
-  border: 1px solid #240370;
+  border: 1px solid white;
   padding: 0.5rem 1.5rem;
   border-radius: 4px;
   cursor: pointer;
@@ -63,8 +67,9 @@ button {
 
 button:hover,
 button:active {
-  background-color: #33059e;
-  border-color: #33059e;
+  background-color: white;
+  border-color: #97BF1C;
+  color: #97BF1C;
 }
 
 button:disabled,
@@ -74,6 +79,16 @@ button:disabled:active {
   color: #292929;
   border-color: #ccc;
   cursor: not-allowed;
+}
+
+.form {
+  width: 300px;
+  margin: 3rem;
+}
+
+.form a {
+  color: #97BF1C;
+  text-align: center;
 }
 
 .form-actions {
