+++ src/store/index.js
@@ -1,11 +1,8 @@
 /* eslint-disable */
 import { configureStore } from '@reduxjs/toolkit';
-import loginSlice from './slices/loginSlice';
 
 const store = configureStore({
- reducer: {
-   login: loginSlice.reducer,
- },
+ reducer: {},
 });
 
 export default store;
